backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
pymysql==1.1.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
scikit-learn==1.4.2
pydantic==2.6.4
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9